requirements.txt
transformers == 4.57.1
torch == 2.9.0
datasets==4.2.0
python-box==7.3.2
-e.